third_party/WebKit/LayoutTests/crypto/sign-verify.html

<!DOCTYPE html>
<html>
<head>
<script src="../fast/js/resources/js-test-pre.js"></script>
<script src="resources/common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<script>
description("Tests cypto.subtle.sign and crypto.subtle.verify");

jsTestIsAsync = true;

importTestKeys().then(function(importedKeys) {
    keys = importedKeys;

    hmacSha1 = {name: 'hmac', hash: {name: 'sha-1'}};

    data = asciiToArrayBuffer("hello");

    // Pass invalid signature parameters to verify()
    shouldThrow("crypto.subtle.verify(hmacSha1, keys.hmacSha1, null, data)");
    shouldThrow("crypto.subtle.verify(hmacSha1, keys.hmacSha1, 'a', data)");
    shouldThrow("crypto.subtle.verify(hmacSha1, keys.hmacSha1, [], data)");

    // Operation does not support signing.
    shouldThrow("crypto.subtle.sign({name: 'sha-1'}, keys.hmacSha1, data)");

    // Key's algorithm must match.
    shouldThrow("crypto.subtle.sign({name: 'hmac', hash: {name: 'sha-256'}}, keys.hmacSha1, data)");

    // ---------------------------------------------------
    // HMAC normalization failures (HmacParams)
    // ---------------------------------------------------
    shouldThrow("crypto.subtle.sign({name: 'hmac'}, keys.hmacSha1, data)");
    shouldThrow("crypto.subtle.sign({name: 'hmac', hash: 3}, keys.hmacSha1, data)");
    shouldThrow("crypto.subtle.sign({name: 'hmac', hash: null}, keys.hmacSha1, data)");
    shouldThrow("crypto.subtle.sign({name: 'hmac', hash: {}}, keys.hmacSha1, data)");
    shouldThrow("crypto.subtle.sign({name: 'hmac', hash: {name: 'foo'}}, keys.hmacSha1, data)");
    shouldThrow("crypto.subtle.sign({name: 'hmac', hash: {name: 'AES-CBC'}}, keys.hmacSha1, data)");

    // ---------------------------------------------------
    // RSASSA-PKCS1-v1_5 normalization failures (RsaSsaParams)
    // ---------------------------------------------------
    shouldThrow("crypto.subtle.sign({name: 'RSASSA-PKCS1-v1_5'}, keys.rsaSsaSha1, data)");
    shouldThrow("crypto.subtle.sign({name: 'RSASSA-PKCS1-v1_5', hash: 3}, keys.rsaSsaSha1, data)");
    shouldThrow("crypto.subtle.sign({name: 'RSASSA-PKCS1-v1_5', hash: null}, keys.rsaSsaSha1, data)");
    shouldThrow("crypto.subtle.sign({name: 'RSASSA-PKCS1-v1_5', hash: {}}, keys.rsaSsaSha1, data)");
    shouldThrow("crypto.subtle.sign({name: 'RSASSA-PKCS1-v1_5', hash: {name: 'foo'}}, keys.rsaSsaSha1, data)");
    shouldThrow("crypto.subtle.sign({name: 'RSASSA-PKCS1-v1_5', hash: {name: 'AES-CBC'}}, keys.rsaSsaSha1, data)");

    return crypto.subtle.sign(hmacSha1, keys.hmacSha1, data);
}).then(function(result) {
    signResult = result;
    shouldBe("signResult.byteLength", "17");

    expectedSignature = asciiToArrayBuffer("signed HMAC:hello");
    return crypto.subtle.verify(hmacSha1, keys.hmacSha1, expectedSignature, data);
}).then(function(result) {
    verifyResult = result;
    shouldBe("verifyResult", "true");

    return crypto.subtle.verify(hmacSha1, keys.hmacSha1, asciiToArrayBuffer("badsignature"), data);
}).then(function(result) {
    verifyResult = result;
    shouldBe("verifyResult", "false");
}).then(finishJSTest, failAndFinishJSTest);

</script>

<script src="../fast/js/resources/js-test-post.js"></script>
</body>
